requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
ao3-api==2.3.1
attrs==25.4.0
beautifulsoup4==4.14.3
boto3==1.42.60
botocore==1.42.60
certifi==2026.2.25
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
jmespath==1.1.0
lxml==6.0.2
multidict==6.7.1
propcache==0.4.1
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
